backend/requirements.txt
 fastapi
uvicorn
langchain
langchain-openai
langchain-google-genai
langchain-community
duckduckgo-search
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
requests
